refactor: io 패키지 view로 이름 변경
