+++ src/main/java/racingcar/view/Printer.java
@@ -1,4 +1,4 @@
-package racingcar.io;
+package racingcar.view;
 
 import racingcar.Game;
